requirements.txt
python-magic-bin==0.4.14
qrcode==7.4.2
wxPython==4.2.0